Create README.md
The system should keep track of peoples credit scores and dues. The system is similar to the real cibil system with an enhancement (Here the related person can view his status). The system should feature: admin login ,Cibil associates login and individual login. Here cibil associates are banks or companies who want to report faulty members. They may send faulty member data and this is passed on to the cibil admin. The cibil admin can view the data and approve it to be added after inspection. The system also consists of member login for which a member first needs to register. He can then check if he is listed in the black list and for what. The member may then apply for the removal along with the receipt of fine paid or other documents of proved good behavior.